added std-style.css
std-style.css has been added in order to keep standard
styles across repositories without the need for
fetching external resources.  This does not scale
and eventually the std-style.css may become a single
central file in the root directory which can be fetched
from the development enviroment, however, for now, while
files and folders are constantly changing and the website
is not one repo but a decentralized amalgamation of many,
this seems the best approach.